[Spark load] build global dict for spark load
